refactor quests
