+++ client/apps/game/src/hooks/helpers/use-buildings.tsx
@@ -1,7 +1,7 @@
 import { configManager } from "@/dojo/setup";
 import { useDojo } from "@/hooks/context/dojo-context";
-import { ResourceIdToMiningType } from "@/ui/utils/utils";
-import { Building, BuildingType, ResourcesIds } from "@bibliothecadao/eternum";
+import { getEntityIdFromKeys, ResourceIdToMiningType } from "@/ui/utils/utils";
+import { Building, BuildingType, ID, ResourcesIds } from "@bibliothecadao/eternum";
 import { useEntityQuery } from "@dojoengine/react";
 import { getComponentValue, Has, HasValue } from "@dojoengine/recs";
 import { useMemo } from "react";
@@ -54,4 +54,27 @@ export const useBuildings = (outerCol: number, outerRow: number) => {
   }, [buildingEntities]);
 
   return buildings as Building[];
+};
+
+export const useBuildingQuantities = (structureEntityId: ID | undefined) => {
+  const {
+    setup: {
+      components: { BuildingQuantityv2 },
+    },
+  } = useDojo();
+  const entityUpdate = useEntityQuery([HasValue(BuildingQuantityv2, { entity_id: structureEntityId || 0 })]);
+
+  const getBuildingQuantity = (buildingType: BuildingType) =>
+    getComponentValue(BuildingQuantityv2, getEntityIdFromKeys([BigInt(structureEntityId || 0), BigInt(buildingType)]))
+      ?.value || 0;
+
+  return useMemo(
+    () => ({
+      food: getBuildingQuantity(BuildingType.Farm) + getBuildingQuantity(BuildingType.FishingVillage),
+      resource: getBuildingQuantity(BuildingType.Resource),
+      workersHut: getBuildingQuantity(BuildingType.WorkersHut),
+      markets: getBuildingQuantity(BuildingType.Market),
+    }),
+    [structureEntityId, entityUpdate],
+  );
 };
