+++ client/apps/game/src/ui/components/military/army-chip.tsx
@@ -4,7 +4,6 @@ import { ReactComponent as Swap } from "@/assets/icons/common/swap.svg";
 import { ReactComponent as Compass } from "@/assets/icons/compass.svg";
 import { useDojo } from "@/hooks/context/dojo-context";
 import { useArmiesAtPosition } from "@/hooks/helpers/use-armies";
-import { armyHasTroops } from "@/hooks/helpers/use-quests";
 import useUIStore from "@/hooks/store/use-ui-store";
 import useNextBlockTimestamp from "@/hooks/use-next-block-timestamp";
 import { Position as PositionInterface } from "@/types/position";
@@ -15,6 +14,7 @@ import { Exchange } from "@/ui/components/structures/worldmap/structure-card";
 import { ArmyCapacity } from "@/ui/elements/army-capacity";
 import Button from "@/ui/elements/button";
 import { StaminaResource } from "@/ui/elements/stamina-resource";
+import { armyHasTroops } from "@/utils/army";
 import { ArmyInfo, BattleManager, Position } from "@bibliothecadao/eternum";
 import { LucideArrowRight } from "lucide-react";
 import React, { Dispatch, SetStateAction, useMemo, useState } from "react";
